requirements.txt
beautifulsoup4==4.12.3
geopy==2.2.0
matplotlib==3.6.3
pandas==1.5.3
python-Levenshtein==0.25.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.1
seaborn==0.13.2
statsmodels==0.14.0
numpy<2.0